refactor: decompose AnthropicLlmAdapter.cs into single-responsibility modules (closes #857)
Extract from the 1053-line monolith into 6 focused modules:
- AnthropicDtos.cs: CallSummaryStat DTO
- AnthropicRequestBuilders.cs: BuildMessagesRequest, AttachTool
- AnthropicResponseParsers.cs: all parse methods, NormalizeStatName, PadDialogueOptionsToThree
- AnthropicResponseImprover.cs: ApplyImprovementAsync, StripImprovementEvaluation
- AnthropicDebugLogger.cs: debug logging + token stats tracking
- AnthropicOverlayApplier.cs: ApplyHorninessOverlayAsync

AnthropicLlmAdapter.cs now orchestrates via delegation. Backward-compat
static methods (ParseDialogueOptions, ParseOpponentResponse) preserved
for existing test references.

Test results: identical to baseline (821 pass, 62 pre-existing failures, 9 skipped).
